# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "basecast-get-data",
    "description": "Serves the basecast lake, the processed tables and the pipeline runs to basecast-app. Every route but /health needs `Authorization: Bearer <API_TOKEN>`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tables": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "List Tables",
        "operationId": "list_tables_tables_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{name}": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "Table Detail",
        "operationId": "table_detail_tables__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,62}$",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{name}/rows": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "Table Rows",
        "operationId": "table_rows_tables__name__rows_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,62}$",
              "title": "Name"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Column to sort by",
              "title": "Sort"
            },
            "description": "Column to sort by"
          },
          {
            "name": "desc",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Desc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "column:op:value, repeatable. Ops: eq ne contains starts in gte lte gt lt between null notnull; `in` and `between` join values with U+001F",
              "default": [],
              "title": "Filter"
            },
            "description": "column:op:value, repeatable. Ops: eq ne contains starts in gte lte gt lt between null notnull; `in` and `between` join values with U+001F"
          },
          {
            "name": "with_summary",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also return the total (exact when it fits in 6 s)",
              "default": false,
              "title": "With Summary"
            },
            "description": "Also return the total (exact when it fits in 6 s)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{name}/lineage": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "Table Lineage",
        "description": "The raw files that fed a table (`lake_processed`); SQL datasets have none of their own.",
        "operationId": "table_lineage_tables__name__lineage_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,62}$",
              "title": "Name"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GridColumn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number",
              "date",
              "boolean",
              "json",
              "geometry"
            ],
            "title": "Type"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "title": "GridColumn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LineagePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LineageRow"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "LineagePage"
      },
      "LineageResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/LineagePage"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "LineageResponse"
      },
      "LineageRow": {
        "properties": {
          "raw_key": {
            "type": "string",
            "title": "Raw Key"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "processed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Processed At"
          }
        },
        "type": "object",
        "required": [
          "raw_key",
          "sha256",
          "version",
          "processed_at"
        ],
        "title": "LineageRow"
      },
...
      "Meta": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data_as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data As Of"
          },
          "model_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Run Id"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": false
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "generated_at"
        ],
        "title": "Meta"
      },
...
      "RowsPage": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/GridColumn"
            },
            "type": "array",
            "title": "Columns"
          },
          "rows": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "type": "array",
            "title": "Rows"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "total_estimated": {
            "type": "boolean",
            "title": "Total Estimated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "columns",
          "rows",
          "offset",
          "limit"
        ],
        "title": "RowsPage",
        "description": "One block of rows. `total` comes only when the request asks for it (`with_summary=true`)."
      },
      "RowsResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/RowsPage"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "RowsResponse"
      },
...
      "TableColumn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number",
              "date",
              "boolean",
              "json",
              "geometry"
            ],
            "title": "Type"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "nullable": {
            "type": "boolean",
            "title": "Nullable"
          },
          "is_key": {
            "type": "boolean",
            "title": "Is Key",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "position",
          "nullable"
        ],
        "title": "TableColumn"
      },
      "TableDetail": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "engine": {
            "type": "string",
            "enum": [
              "postgres",
              "bigquery"
            ],
            "title": "Engine"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "kind": {
            "type": "string",
            "enum": [
              "dataset",
              "system",
              "unregistered"
            ],
            "title": "Kind"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "declared": {
            "type": "boolean",
            "title": "Declared"
          },
          "loaded": {
            "type": "boolean",
            "title": "Loaded"
          },
          "rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "rows_estimated": {
            "type": "boolean",
            "title": "Rows Estimated",
            "default": false
          },
          "bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/TableColumn"
            },
            "type": "array",
            "title": "Columns"
          },
          "key_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Columns",
            "default": []
          },
          "partition_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partition Field"
          },
          "cluster_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cluster Fields",
            "default": []
          },
          "derived": {
            "type": "boolean",
            "title": "Derived",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "engine",
          "location",
          "kind",
          "sources",
          "declared",
          "loaded",
          "columns"
        ],
        "title": "TableDetail"
      },
      "TableResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/TableDetail"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "TableResponse"
      },
      "TableSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "engine": {
            "type": "string",
            "enum": [
              "postgres",
              "bigquery"
            ],
            "title": "Engine"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "kind": {
            "type": "string",
            "enum": [
              "dataset",
              "system",
              "unregistered"
            ],
            "title": "Kind"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "declared": {
            "type": "boolean",
            "title": "Declared"
          },
          "loaded": {
            "type": "boolean",
            "title": "Loaded"
          },
          "rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "rows_estimated": {
            "type": "boolean",
            "title": "Rows Estimated",
            "default": false
          },
          "bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "engine",
          "location",
          "kind",
          "sources",
          "declared",
          "loaded"
        ],
        "title": "TableSummary"
      },
      "TablesData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TableSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "TablesData"
      },
      "TablesResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/TablesData"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "TablesResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}